refactor(renderer): move _new_design group settings into groups/ and applets/
Final slice of dissolving _new_design/. The group-settings cluster moves
to its ontological homes:
- groups/elements/settings/ — group-settings + general/tools/inactive-tools
  /danger-zone/my-profile settings (+ shared tool-settings styles/utils)
- groups/elements/members/  — group-member-list, agent-permission-button
- groups/elements/invite/   — invite-people-dialog, select-group
- applets/elements/         — applet/base-applet/abandoned-applet settings cards

The _new_design/ folder — named for a redesign era rather than for what it
contained — is now empty and removed. Pure file moves + import-path
updates; no behaviour change.
