+++ src/renderer/src/applets/elements/base-applet-settings-card.ts
@@ -12,20 +12,20 @@ import '@shoelace-style/shoelace/dist/components/icon/icon.js';
 import '@shoelace-style/shoelace/dist/components/tooltip/tooltip.js';
 import '@shoelace-style/shoelace/dist/components/button/button.js';
 
-import '../../../ui/copy-hash';
-import '../../../ui/moss-mini-button.js';
-import { deprecateIcon, devIcon } from '../../../ui/icons.js';
-import { deprecateTool, undeprecateTool } from './tool-settings-utils.js';
+import '../../ui/copy-hash';
+import '../../ui/moss-mini-button.js';
+import { deprecateIcon, devIcon } from '../../ui/icons.js';
+import { deprecateTool, undeprecateTool } from '../../groups/elements/settings/tool-settings-utils.js';
 
 import { ALWAYS_ONLINE_TAG, Applet, GroupAppletsMetaData } from '@theweave/group-client';
 
 import { StoreSubscriber, lazyLoadAndPoll } from '@holochain-open-dev/stores';
-import { mossStoreContext } from '../../../context.js';
-import { MossStore } from '../../../moss-store.js';
-import { groupStoreContext } from '../../../groups/context.js';
-import { GroupStore } from '../../../groups/group-store.js';
-import { mossStyles } from '../../../shared-styles.js';
-import { toolSettingsStyles } from './tool-settings-styles.js';
+import { mossStoreContext } from '../../context.js';
+import { MossStore } from '../../moss-store.js';
+import { groupStoreContext } from '../../groups/context.js';
+import { GroupStore } from '../../groups/group-store.js';
+import { mossStyles } from '../../shared-styles.js';
+import { toolSettingsStyles } from '../../groups/elements/settings/tool-settings-styles.js';
 
 /**
  * Base class for applet settings cards.
