+++ src/renderer/src/groups/elements/group-container.ts
@@ -15,13 +15,13 @@ import './applet-main-views.js';
 import '../../app/navigation/group-area-sidebar.js';
 import './group-home.js';
 import '../../self/profile/moss-profile-prompt.js';
-import '../../elements/_new_design/group-settings/my-profile-settings.js';
+import './settings/my-profile-settings.js';
 import { wrapPathInSvg } from '@holochain-open-dev/elements';
 import { mdiPowerPlugOff } from '@mdi/js';
 import { StoreSubscriber } from '@holochain-open-dev/stores';
 import { AppletHash } from '@theweave/api';
 import { GroupHome } from './group-home.js';
-import { MyProfileSettings } from '../../elements/_new_design/group-settings/my-profile-settings.js';
+import { MyProfileSettings } from './settings/my-profile-settings.js';
 import { MossDialog } from '../../ui/moss-dialog.js';
 import '../../ui/moss-dialog.js';
