+++ src/renderer/src/groups/elements/group-home.ts
@@ -47,11 +47,11 @@ import './edit-custom-group-view.js';
 import '../../elements/reusable/tab-group.js';
 import './foyer-stream.js';
 import './agent-permission.js';
-import '../../elements/_new_design/group-settings.js';
+import './settings/group-settings.js';
 import '../../ui/moss-dialog.js';
 import '../../self/profile/moss-profile-detail.js';
 import '../../ui/copy-hash.js';
-import '../../elements/_new_design/invite-people-dialog.js';
+import './invite/invite-people-dialog.js';
 
 import { groupStoreContext } from '../context.js';
 import { GroupStore } from '../group-store.js';
@@ -67,7 +67,7 @@ import { Applet, AppletAgent } from '../../../../../shared/group-client/dist/ind
 import { markdownParseSafe } from '../../utils.js';
 import { unsafeHTML } from 'lit/directives/unsafe-html.js';
 import { appIdFromAppletHash } from '@theweave/utils';
-import { GroupSettings } from '../../elements/_new_design/group-settings.js';
+import { GroupSettings } from './settings/group-settings.js';
 import { MossDialog } from '../../ui/moss-dialog.js';
 import { editIcon, closeIcon, saveIcon, personPlusIcon } from '../../ui/icons.js';
 import yaml from "js-yaml";
