+++ src/renderer/src/groups/elements/settings/inactive-tools.ts
@@ -8,8 +8,8 @@ import { css, html, LitElement } from 'lit';
 import { localized, msg, str } from '@lit/localize';
 import { mossStoreContext } from '../../../context';
 import { MossStore } from '../../../moss-store';
-import { groupStoreContext } from '../../../groups/context';
-import { GroupStore } from '../../../groups/group-store';
+import { groupStoreContext } from '../../context';
+import { GroupStore } from '../../group-store';
 import { AgentPubKey, encodeHashToBase64 } from '@holochain/client';
 import { notify, notifyError } from '@holochain-open-dev/elements';
 import { mossStyles } from '../../../shared-styles';
