+++ src/renderer/src/groups/elements/settings/tool-settings-styles.ts
